% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpVect-structs.R
\name{as.list.cpVect}
\alias{as.list.cpVect}
\title{Unpack cpVect struct to a list}
\usage{
\method{as.list}{cpVect}(x, ...)
}
\arguments{
\item{x}{\code{cpVect *}}

\item{...}{other arguments ignored}
}
\description{
Unpack cpVect struct to a list
}
