% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6.R
\name{Chipmunk}
\alias{Chipmunk}
\title{Chipmunk class}
\description{
Chipmunk class

Chipmunk class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Chipmunk$new()}}
\item \href{#method-add_static_segment}{\code{Chipmunk$add_static_segment()}}
\item \href{#method-get_static_segments}{\code{Chipmunk$get_static_segments()}}
\item \href{#method-add_circle}{\code{Chipmunk$add_circle()}}
\item \href{#method-get_circles}{\code{Chipmunk$get_circles()}}
\item \href{#method-add_box}{\code{Chipmunk$add_box()}}
\item \href{#method-get_boxes}{\code{Chipmunk$get_boxes()}}
\item \href{#method-get_boxes_as_polygons}{\code{Chipmunk$get_boxes_as_polygons()}}
\item \href{#method-remove_circle}{\code{Chipmunk$remove_circle()}}
\item \href{#method-advance}{\code{Chipmunk$advance()}}
\item \href{#method-clone}{\code{Chipmunk$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize the simulation space
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Chipmunk$new(gravity = cpv(0, -100), time_step = 0.01)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{gravity}}{vector object indication direction and strength
of gravity. Default: \code{cpv(0, -100)}}

\item{\code{time_step}}{simulation step size. smaller is more accurate.
       Default: 0.01
Add a static segment}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_static_segment"></a>}}
\if{latex}{\out{\hypertarget{method-add_static_segment}{}}}
\subsection{Method \code{add_static_segment()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Chipmunk$add_static_segment(x1, y1, x2, y2, friction = 1, elasticity = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x1, y1, x2, y2}}{segment end points}

\item{\code{friction}}{friction along ground. default 1.}

\item{\code{elasticity}}{default: 0 (no bounce)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_static_segments"></a>}}
\if{latex}{\out{\hypertarget{method-get_static_segments}{}}}
\subsection{Method \code{get_static_segments()}}{
Get the data.frame of all current segments
Add a circular body to the space
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Chipmunk$get_static_segments()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_circle"></a>}}
\if{latex}{\out{\hypertarget{method-add_circle}{}}}
\subsection{Method \code{add_circle()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Chipmunk$add_circle(
  x,
  y,
  vx = 0,
  vy = 0,
  radius = 1,
  mass = 1,
  friction = 0.7,
  elasticity = 0
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x, y}}{initial body location}

\item{\code{vx, vy}}{initial body velocity}

\item{\code{radius}}{radius of body. default: 1}

\item{\code{mass}}{mass of body. default: 1}

\item{\code{friction}}{default: 0.7}

\item{\code{elasticity}}{default: 0  (no bounce).  Valid range [0, 1]}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_circles"></a>}}
\if{latex}{\out{\hypertarget{method-get_circles}{}}}
\subsection{Method \code{get_circles()}}{
Get the location of all the bodies
Add a box body to the space
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Chipmunk$get_circles()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_box"></a>}}
\if{latex}{\out{\hypertarget{method-add_box}{}}}
\subsection{Method \code{add_box()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Chipmunk$add_box(
  x,
  y,
  vx = 0,
  vy = 0,
  width = 1.73,
  height = 1,
  angle = 0,
  radius = 0.05,
  mass = 1,
  friction = 0.7,
  elasticity = 0
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x, y}}{initial body location}

\item{\code{vx, vy}}{initial body velocity}

\item{\code{width, height}}{body width and height}

\item{\code{angle}}{rotation angle in degrees. default 0}

\item{\code{radius}}{radius of rounded corner}

\item{\code{mass}}{mass of body. default: 1}

\item{\code{friction}}{default: 0.7}

\item{\code{elasticity}}{default: 0  (no bounce).  Valid range [0, 1]}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_boxes"></a>}}
\if{latex}{\out{\hypertarget{method-get_boxes}{}}}
\subsection{Method \code{get_boxes()}}{
Get the location of all the bodies
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Chipmunk$get_boxes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a data.frame with idx, width, height and rotation angle (radians)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_boxes_as_polygons"></a>}}
\if{latex}{\out{\hypertarget{method-get_boxes_as_polygons}{}}}
\subsection{Method \code{get_boxes_as_polygons()}}{
Get the location of all the boxes as a list of corner vertices
Remove a circle from the simulation

To speed up the simulation process, you could test bodies for
location and remove those you are no longer interested in.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Chipmunk$get_boxes_as_polygons()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-remove_circle"></a>}}
\if{latex}{\out{\hypertarget{method-remove_circle}{}}}
\subsection{Method \code{remove_circle()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Chipmunk$remove_circle(idx)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{idx}}{index of circle to remove
Advance the simulation by N timesteps}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-advance"></a>}}
\if{latex}{\out{\hypertarget{method-advance}{}}}
\subsection{Method \code{advance()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Chipmunk$advance(N = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{N}}{number of time steps to advance}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Chipmunk$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
