% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chipmunk.R
\name{cpAreaForCircle}
\alias{cpAreaForCircle}
\title{Calculate area of a hollow circle. @c r1 and @c r2 are the inner and outer
diameters. A solid circle has an inner diameter of 0.}
\usage{
cpAreaForCircle(r1, r2)
}
\arguments{
\item{r1}{[\code{cpFloat}]}

\item{r2}{[\code{cpFloat}]}
}
\value{
[\code{cpFloat}]
}
\description{
Calculate area of a hollow circle. @c r1 and @c r2 are the inner and outer
diameters. A solid circle has an inner diameter of 0.
}
\details{
C function prototype: \code{CP_EXPORT cpFloat cpAreaForCircle(cpFloat r1, cpFloat r2);}
}
