% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chipmunk.R
\name{cpAreaForPoly}
\alias{cpAreaForPoly}
\title{Calculate the signed area of a polygon. A Clockwise winding gives positive area.
This is probably backwards from what you expect, but matches Chipmunk's the
winding for poly shapes.}
\usage{
cpAreaForPoly(count, verts, radius)
}
\arguments{
\item{count}{[\code{int}]}

\item{verts}{[\code{cpVect *}]}

\item{radius}{[\code{cpFloat}]}
}
\value{
[\code{cpFloat}]
}
\description{
Calculate the signed area of a polygon. A Clockwise winding gives positive area.
This is probably backwards from what you expect, but matches Chipmunk's the
winding for poly shapes.
}
\details{
C function prototype: \code{CP_EXPORT cpFloat cpAreaForPoly(const int count, const cpVect *verts, cpFloat radius);}
}
