% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chipmunk.R
\name{cpAreaForSegment}
\alias{cpAreaForSegment}
\title{Calculate the area of a fattened (capsule shaped) line segment.}
\usage{
cpAreaForSegment(a, b, radius)
}
\arguments{
\item{a}{[\code{cpVect *}]}

\item{b}{[\code{cpVect *}]}

\item{radius}{[\code{cpFloat}]}
}
\value{
[\code{cpFloat}]
}
\description{
Calculate the area of a fattened (capsule shaped) line segment.
}
\details{
C function prototype: \code{CP_EXPORT cpFloat cpAreaForSegment(cpVect a, cpVect b, cpFloat radius);}
}
