% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpBody.R
\name{cpBodyGetAngle}
\alias{cpBodyGetAngle}
\title{Get the angle of the body.}
\usage{
cpBodyGetAngle(body)
}
\arguments{
\item{body}{[\code{cpBody *}]}
}
\value{
[\code{cpFloat}]
}
\description{
Get the angle of the body.
}
\details{
C function prototype: \code{CP_EXPORT cpFloat cpBodyGetAngle(const cpBody *body);}
}
