% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpBody.R
\name{cpBodyGetAngularVelocity}
\alias{cpBodyGetAngularVelocity}
\title{Get the angular velocity of the body.}
\usage{
cpBodyGetAngularVelocity(body)
}
\arguments{
\item{body}{[\code{cpBody *}]}
}
\value{
[\code{cpFloat}]
}
\description{
Get the angular velocity of the body.
}
\details{
C function prototype: \code{CP_EXPORT cpFloat cpBodyGetAngularVelocity(const cpBody *body);}
}
