% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpBody.R
\name{cpBodyGetCenterOfGravity}
\alias{cpBodyGetCenterOfGravity}
\title{Get the offset of the center of gravity in body local coordinates.}
\usage{
cpBodyGetCenterOfGravity(body)
}
\arguments{
\item{body}{[\code{cpBody *}]}
}
\value{
[\code{cpVect *}]
}
\description{
Get the offset of the center of gravity in body local coordinates.
}
\details{
C function prototype: \code{CP_EXPORT cpVect cpBodyGetCenterOfGravity(const cpBody *body);}
}
