% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpBody.R
\name{cpBodyGetForce}
\alias{cpBodyGetForce}
\title{Get the force applied to the body for the next time step.}
\usage{
cpBodyGetForce(body)
}
\arguments{
\item{body}{[\code{cpBody *}]}
}
\value{
[\code{cpVect *}]
}
\description{
Get the force applied to the body for the next time step.
}
\details{
C function prototype: \code{CP_EXPORT cpVect cpBodyGetForce(const cpBody *body);}
}
