% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpBody.R
\name{cpBodyGetPosition}
\alias{cpBodyGetPosition}
\title{Set the position of a body.}
\usage{
cpBodyGetPosition(body)
}
\arguments{
\item{body}{[\code{cpBody *}]}
}
\value{
[\code{cpVect *}]
}
\description{
Set the position of a body.
}
\details{
C function prototype: \code{CP_EXPORT cpVect cpBodyGetPosition(const cpBody *body);}
}
