% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpBody.R
\name{cpBodyGetRotation}
\alias{cpBodyGetRotation}
\title{Get the rotation vector of the body. (The x basis vector of it's transform.)}
\usage{
cpBodyGetRotation(body)
}
\arguments{
\item{body}{[\code{cpBody *}]}
}
\value{
[\code{cpVect *}]
}
\description{
Get the rotation vector of the body. (The x basis vector of it's transform.)
}
\details{
C function prototype: \code{CP_EXPORT cpVect cpBodyGetRotation(const cpBody *body);}
}
