% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpBody.R
\name{cpBodyGetTorque}
\alias{cpBodyGetTorque}
\title{Get the torque applied to the body for the next time step.}
\usage{
cpBodyGetTorque(body)
}
\arguments{
\item{body}{[\code{cpBody *}]}
}
\value{
[\code{cpFloat}]
}
\description{
Get the torque applied to the body for the next time step.
}
\details{
C function prototype: \code{CP_EXPORT cpFloat cpBodyGetTorque(const cpBody *body);}
}
