% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpBody.R
\name{cpBodyGetVelocity}
\alias{cpBodyGetVelocity}
\title{Get the velocity of the body.}
\usage{
cpBodyGetVelocity(body)
}
\arguments{
\item{body}{[\code{cpBody *}]}
}
\value{
[\code{cpVect *}]
}
\description{
Get the velocity of the body.
}
\details{
C function prototype: \code{CP_EXPORT cpVect cpBodyGetVelocity(const cpBody *body);}
}
