% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpBody.R
\name{cpBodyNew}
\alias{cpBodyNew}
\title{Allocate and initialize a cpBody.}
\usage{
cpBodyNew(mass, moment)
}
\arguments{
\item{mass}{[\code{cpFloat}]}

\item{moment}{[\code{cpFloat}]}
}
\value{
[\code{cpBody *}]
}
\description{
Allocate and initialize a cpBody.
}
\details{
C function prototype: \code{CP_EXPORT cpBody* cpBodyNew(cpFloat mass, cpFloat moment);}
}
