% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpBody.R
\name{cpBodySetAngle}
\alias{cpBodySetAngle}
\title{Set the angle of a body.}
\usage{
cpBodySetAngle(body, a)
}
\arguments{
\item{body}{[\code{cpBody *}]}

\item{a}{[\code{cpFloat}]}
}
\description{
Set the angle of a body.
}
\details{
C function prototype: \code{CP_EXPORT void cpBodySetAngle(cpBody *body, cpFloat a);}
}
