% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpBody.R
\name{cpBodySetAngularVelocity}
\alias{cpBodySetAngularVelocity}
\title{Set the angular velocity of the body.}
\usage{
cpBodySetAngularVelocity(body, angularVelocity)
}
\arguments{
\item{body}{[\code{cpBody *}]}

\item{angularVelocity}{[\code{cpFloat}]}
}
\description{
Set the angular velocity of the body.
}
\details{
C function prototype: \code{CP_EXPORT void cpBodySetAngularVelocity(cpBody *body, cpFloat angularVelocity);}
}
