% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpBody.R
\name{cpBodySetPosition}
\alias{cpBodySetPosition}
\title{Set the position of the body.}
\usage{
cpBodySetPosition(body, pos)
}
\arguments{
\item{body}{[\code{cpBody *}]}

\item{pos}{[\code{cpVect *}]}
}
\description{
Set the position of the body.
}
\details{
C function prototype: \code{CP_EXPORT void cpBodySetPosition(cpBody *body, cpVect pos);}
}
