% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpBody.R
\name{cpBodySetTorque}
\alias{cpBodySetTorque}
\title{Set the torque applied to the body for the next time step.}
\usage{
cpBodySetTorque(body, torque)
}
\arguments{
\item{body}{[\code{cpBody *}]}

\item{torque}{[\code{cpFloat}]}
}
\description{
Set the torque applied to the body for the next time step.
}
\details{
C function prototype: \code{CP_EXPORT void cpBodySetTorque(cpBody *body, cpFloat torque);}
}
