% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpBody.R
\name{cpBodySetVelocity}
\alias{cpBodySetVelocity}
\title{Set the velocity of the body.}
\usage{
cpBodySetVelocity(body, velocity)
}
\arguments{
\item{body}{[\code{cpBody *}]}

\item{velocity}{[\code{cpVect *}]}
}
\description{
Set the velocity of the body.
}
\details{
C function prototype: \code{CP_EXPORT void cpBodySetVelocity(cpBody *body, cpVect velocity);}
}
