% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chipmunk.R
\name{cpCentroidForPoly}
\alias{cpCentroidForPoly}
\title{Calculate the natural centroid of a polygon.}
\usage{
cpCentroidForPoly(count, verts)
}
\arguments{
\item{count}{[\code{int}]}

\item{verts}{[\code{cpVect *}]}
}
\value{
[\code{cpVect *}]
}
\description{
Calculate the natural centroid of a polygon.
}
\details{
C function prototype: \code{CP_EXPORT cpVect cpCentroidForPoly(const int count, const cpVect *verts);}
}
