% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpConstraint.R
\name{cpConstraintSetErrorBias}
\alias{cpConstraintSetErrorBias}
\title{Set rate at which joint error is corrected. Defaults to pow(1.0 - 0.1, 60.0)
meaning that it will correct 10\% of the error every 1/60th of a second.}
\usage{
cpConstraintSetErrorBias(constraint, errorBias)
}
\arguments{
\item{constraint}{[\code{cpConstraint *}]}

\item{errorBias}{[\code{cpFloat}]}
}
\description{
Set rate at which joint error is corrected. Defaults to pow(1.0 - 0.1, 60.0)
meaning that it will correct 10\% of the error every 1/60th of a second.
}
\details{
C function prototype: \code{CP_EXPORT void cpConstraintSetErrorBias(cpConstraint *constraint, cpFloat errorBias);}
}
