% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpConstraint.R
\name{cpConstraintSetMaxBias}
\alias{cpConstraintSetMaxBias}
\title{Set the maximum rate at which joint error is corrected. (defaults to INFINITY)}
\usage{
cpConstraintSetMaxBias(constraint, maxBias)
}
\arguments{
\item{constraint}{[\code{cpConstraint *}]}

\item{maxBias}{[\code{cpFloat}]}
}
\description{
Set the maximum rate at which joint error is corrected. (defaults to INFINITY)
}
\details{
C function prototype: \code{CP_EXPORT void cpConstraintSetMaxBias(cpConstraint *constraint, cpFloat maxBias);}
}
