% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpConstraint.R
\name{cpConstraintSetMaxForce}
\alias{cpConstraintSetMaxForce}
\title{Set the maximum force that this constraint is allowed to use. (defaults to
INFINITY)}
\usage{
cpConstraintSetMaxForce(constraint, maxForce)
}
\arguments{
\item{constraint}{[\code{cpConstraint *}]}

\item{maxForce}{[\code{cpFloat}]}
}
\description{
Set the maximum force that this constraint is allowed to use. (defaults to
INFINITY)
}
\details{
C function prototype: \code{CP_EXPORT void cpConstraintSetMaxForce(cpConstraint *constraint, cpFloat maxForce);}
}
