% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpDampedSpring.R
\name{cpDampedSpringNew}
\alias{cpDampedSpringNew}
\title{Allocate and initialize a damped spring.}
\usage{
cpDampedSpringNew(a, b, anchorA, anchorB, restLength, stiffness, damping)
}
\arguments{
\item{a}{[\code{cpBody *}]}

\item{b}{[\code{cpBody *}]}

\item{anchorA}{[\code{cpVect *}]}

\item{anchorB}{[\code{cpVect *}]}

\item{restLength}{[\code{cpFloat}]}

\item{stiffness}{[\code{cpFloat}]}

\item{damping}{[\code{cpFloat}]}
}
\value{
[\code{cpConstraint *}]
}
\description{
Allocate and initialize a damped spring.
}
\details{
C function prototype: \code{CP_EXPORT cpConstraint* cpDampedSpringNew(cpBody *a, cpBody *b, cpVect anchorA, cpVect anchorB, cpFloat restLength, cpFloat stiffness, cpFloat damping);}
}
