% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpDampedSpring.R
\name{cpDampedSpringSetDamping}
\alias{cpDampedSpringSetDamping}
\title{Set the damping of the spring.}
\usage{
cpDampedSpringSetDamping(constraint, damping)
}
\arguments{
\item{constraint}{[\code{cpConstraint *}]}

\item{damping}{[\code{cpFloat}]}
}
\description{
Set the damping of the spring.
}
\details{
C function prototype: \code{CP_EXPORT void cpDampedSpringSetDamping(cpConstraint *constraint, cpFloat damping);}
}
