% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpDampedSpring.R
\name{cpDampedSpringSetRestLength}
\alias{cpDampedSpringSetRestLength}
\title{Set the rest length of the spring.}
\usage{
cpDampedSpringSetRestLength(constraint, restLength)
}
\arguments{
\item{constraint}{[\code{cpConstraint *}]}

\item{restLength}{[\code{cpFloat}]}
}
\description{
Set the rest length of the spring.
}
\details{
C function prototype: \code{CP_EXPORT void cpDampedSpringSetRestLength(cpConstraint *constraint, cpFloat restLength);}
}
