% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpDampedSpring.R
\name{cpDampedSpringSetStiffness}
\alias{cpDampedSpringSetStiffness}
\title{Set the stiffness of the spring in force/distance.}
\usage{
cpDampedSpringSetStiffness(constraint, stiffness)
}
\arguments{
\item{constraint}{[\code{cpConstraint *}]}

\item{stiffness}{[\code{cpFloat}]}
}
\description{
Set the stiffness of the spring in force/distance.
}
\details{
C function prototype: \code{CP_EXPORT void cpDampedSpringSetStiffness(cpConstraint *constraint, cpFloat stiffness);}
}
