% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chipmunk.R
\name{cpMomentForBox}
\alias{cpMomentForBox}
\title{Calculate the moment of inertia for a solid box.}
\usage{
cpMomentForBox(m, width, height)
}
\arguments{
\item{m}{[\code{cpFloat}]}

\item{width}{[\code{cpFloat}]}

\item{height}{[\code{cpFloat}]}
}
\value{
[\code{cpFloat}]
}
\description{
Calculate the moment of inertia for a solid box.
}
\details{
C function prototype: \code{CP_EXPORT cpFloat cpMomentForBox(cpFloat m, cpFloat width, cpFloat height);}
}
