% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chipmunk.R
\name{cpMomentForCircle}
\alias{cpMomentForCircle}
\title{Calculate the moment of inertia for a circle. @c r1 and @c r2 are the inner and
outer diameters. A solid circle has an inner diameter of 0.}
\usage{
cpMomentForCircle(m, r1, r2, offset)
}
\arguments{
\item{m}{[\code{cpFloat}]}

\item{r1}{[\code{cpFloat}]}

\item{r2}{[\code{cpFloat}]}

\item{offset}{[\code{cpVect *}]}
}
\value{
[\code{cpFloat}]
}
\description{
Calculate the moment of inertia for a circle. @c r1 and @c r2 are the inner and
outer diameters. A solid circle has an inner diameter of 0.
}
\details{
C function prototype: \code{CP_EXPORT cpFloat cpMomentForCircle(cpFloat m, cpFloat r1, cpFloat r2, cpVect offset);}
}
