% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chipmunk.R
\name{cpMomentForPoly}
\alias{cpMomentForPoly}
\title{Calculate the moment of inertia for a solid polygon shape assuming it's center
of gravity is at it's centroid. The offset is added to each vertex.}
\usage{
cpMomentForPoly(m, count, verts, offset, radius)
}
\arguments{
\item{m}{[\code{cpFloat}]}

\item{count}{[\code{int}]}

\item{verts}{[\code{cpVect *}]}

\item{offset}{[\code{cpVect *}]}

\item{radius}{[\code{cpFloat}]}
}
\value{
[\code{cpFloat}]
}
\description{
Calculate the moment of inertia for a solid polygon shape assuming it's center
of gravity is at it's centroid. The offset is added to each vertex.
}
\details{
C function prototype: \code{CP_EXPORT cpFloat cpMomentForPoly(cpFloat m, int count, const cpVect *verts, cpVect offset, cpFloat radius);}
}
