% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chipmunk.R
\name{cpMomentForSegment}
\alias{cpMomentForSegment}
\title{Calculate the moment of inertia for a line segment. Beveling radius is not
supported.}
\usage{
cpMomentForSegment(m, a, b, radius)
}
\arguments{
\item{m}{[\code{cpFloat}]}

\item{a}{[\code{cpVect *}]}

\item{b}{[\code{cpVect *}]}

\item{radius}{[\code{cpFloat}]}
}
\value{
[\code{cpFloat}]
}
\description{
Calculate the moment of inertia for a line segment. Beveling radius is not
supported.
}
\details{
C function prototype: \code{CP_EXPORT cpFloat cpMomentForSegment(cpFloat m, cpVect a, cpVect b, cpFloat radius);}
}
