% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpPolyShape.R
\name{cpPolyShapeNew}
\alias{cpPolyShapeNew}
\title{Allocate and initialize a polygon shape with rounded corners. A convex hull will
be created from the vertexes.}
\usage{
cpPolyShapeNew(body, count, verts, transform, radius)
}
\arguments{
\item{body}{[\code{cpBody *}]}

\item{count}{[\code{int}]}

\item{verts}{[\code{cpVect *}]}

\item{transform}{[\code{cpTransform *}]}

\item{radius}{[\code{cpFloat}]}
}
\value{
[\code{cpShape *}]
}
\description{
Allocate and initialize a polygon shape with rounded corners. A convex hull will
be created from the vertexes.
}
\details{
C function prototype: \code{CP_EXPORT cpShape* cpPolyShapeNew(cpBody *body, int count, const cpVect *verts, cpTransform transform, cpFloat radius);}
}
