% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpShape.R
\name{cpSegmentShapeNew}
\alias{cpSegmentShapeNew}
\title{Allocate and initialize a segment shape.}
\usage{
cpSegmentShapeNew(body, a, b, radius)
}
\arguments{
\item{body}{[\code{cpBody *}]}

\item{a}{[\code{cpVect *}]}

\item{b}{[\code{cpVect *}]}

\item{radius}{[\code{cpFloat}]}
}
\value{
[\code{cpShape *}]
}
\description{
Allocate and initialize a segment shape.
}
\details{
C function prototype: \code{CP_EXPORT cpShape* cpSegmentShapeNew(cpBody *body, cpVect a, cpVect b, cpFloat radius);}
}
