% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpShape.R
\name{cpShapeGetArea}
\alias{cpShapeGetArea}
\title{Get the calculated area of this shape.}
\usage{
cpShapeGetArea(shape)
}
\arguments{
\item{shape}{[\code{cpShape *}]}
}
\value{
[\code{cpFloat}]
}
\description{
Get the calculated area of this shape.
}
\details{
C function prototype: \code{CP_EXPORT cpFloat cpShapeGetArea(cpShape *shape);}
}
