% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpShape.R
\name{cpShapeGetCenterOfGravity}
\alias{cpShapeGetCenterOfGravity}
\title{Get the centroid of this shape.}
\usage{
cpShapeGetCenterOfGravity(shape)
}
\arguments{
\item{shape}{[\code{cpShape *}]}
}
\value{
[\code{cpVect *}]
}
\description{
Get the centroid of this shape.
}
\details{
C function prototype: \code{CP_EXPORT cpVect cpShapeGetCenterOfGravity(cpShape *shape);}
}
