% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpShape.R
\name{cpShapeGetMoment}
\alias{cpShapeGetMoment}
\title{Get the calculated moment of inertia for this shape.}
\usage{
cpShapeGetMoment(shape)
}
\arguments{
\item{shape}{[\code{cpShape *}]}
}
\value{
[\code{cpFloat}]
}
\description{
Get the calculated moment of inertia for this shape.
}
\details{
C function prototype: \code{CP_EXPORT cpFloat cpShapeGetMoment(cpShape *shape);}
}
