% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpShape.R
\name{cpShapeSetDensity}
\alias{cpShapeSetDensity}
\title{Set the density of this shape to have Chipmunk calculate mass properties for
you.}
\usage{
cpShapeSetDensity(shape, density)
}
\arguments{
\item{shape}{[\code{cpShape *}]}

\item{density}{[\code{cpFloat}]}
}
\description{
Set the density of this shape to have Chipmunk calculate mass properties for
you.
}
\details{
C function prototype: \code{CP_EXPORT void cpShapeSetDensity(cpShape *shape, cpFloat density);}
}
