% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpShape.R
\name{cpShapeSetElasticity}
\alias{cpShapeSetElasticity}
\title{Set the elasticity of this shape.}
\usage{
cpShapeSetElasticity(shape, elasticity)
}
\arguments{
\item{shape}{[\code{cpShape *}]}

\item{elasticity}{[\code{cpFloat}]}
}
\description{
Set the elasticity of this shape.
}
\details{
C function prototype: \code{CP_EXPORT void cpShapeSetElasticity(cpShape *shape, cpFloat elasticity);}
}
