% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpShape.R
\name{cpShapeSetFriction}
\alias{cpShapeSetFriction}
\title{Set the friction of this shape.}
\usage{
cpShapeSetFriction(shape, friction)
}
\arguments{
\item{shape}{[\code{cpShape *}]}

\item{friction}{[\code{cpFloat}]}
}
\description{
Set the friction of this shape.
}
\details{
C function prototype: \code{CP_EXPORT void cpShapeSetFriction(cpShape *shape, cpFloat friction);}
}
