% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpShape.R
\name{cpShapeSetMass}
\alias{cpShapeSetMass}
\title{Set the mass of this shape to have Chipmunk calculate mass properties for you.}
\usage{
cpShapeSetMass(shape, mass)
}
\arguments{
\item{shape}{[\code{cpShape *}]}

\item{mass}{[\code{cpFloat}]}
}
\description{
Set the mass of this shape to have Chipmunk calculate mass properties for you.
}
\details{
C function prototype: \code{CP_EXPORT void cpShapeSetMass(cpShape *shape, cpFloat mass);}
}
