% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpShape.R
\name{cpShapeSetSensor}
\alias{cpShapeSetSensor}
\title{Set if the shape is a sensor or not.}
\usage{
cpShapeSetSensor(shape, sensor)
}
\arguments{
\item{shape}{[\code{cpShape *}]}

\item{sensor}{[\code{cpBool}]}
}
\description{
Set if the shape is a sensor or not.
}
\details{
C function prototype: \code{CP_EXPORT void cpShapeSetSensor(cpShape *shape, cpBool sensor); }
}
