% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpSpace.R
\name{cpSpaceAddBody}
\alias{cpSpaceAddBody}
\title{Add a rigid body to the simulation.}
\usage{
cpSpaceAddBody(space, body)
}
\arguments{
\item{space}{[\code{cpSpace *}]}

\item{body}{[\code{cpBody *}]}
}
\description{
Add a rigid body to the simulation.
}
\details{
C function prototype: \code{CP_EXPORT void cpSpaceAddBody(cpSpace *space, cpBody *body);}
}
