% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpSpace.R
\name{cpSpaceAddConstraint}
\alias{cpSpaceAddConstraint}
\title{Add a constraint to the simulation.}
\usage{
cpSpaceAddConstraint(space, constraint)
}
\arguments{
\item{space}{[\code{cpSpace *}]}

\item{constraint}{[\code{cpConstraint *}]}
}
\description{
Add a constraint to the simulation.
}
\details{
C function prototype: \code{CP_EXPORT void cpSpaceAddConstraint(cpSpace *space, cpConstraint *constraint);}
}
