% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpSpace.R
\name{cpSpaceAddShape}
\alias{cpSpaceAddShape}
\title{Add a collision shape to the simulation. If the shape is attached to a static
body, it will be added as a static shape.}
\usage{
cpSpaceAddShape(space, shape)
}
\arguments{
\item{space}{[\code{cpSpace *}]}

\item{shape}{[\code{cpShape *}]}
}
\description{
Add a collision shape to the simulation. If the shape is attached to a static
body, it will be added as a static shape.
}
\details{
C function prototype: \code{CP_EXPORT void cpSpaceAddShape(cpSpace *space, cpShape *shape);}
}
