% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpSpace.R
\name{cpSpaceContainsBody}
\alias{cpSpaceContainsBody}
\title{Test if a rigid body has been added to the space.}
\usage{
cpSpaceContainsBody(space, body)
}
\arguments{
\item{space}{[\code{cpSpace *}]}

\item{body}{[\code{cpBody *}]}
}
\value{
[\code{cpBool}]
}
\description{
Test if a rigid body has been added to the space.
}
\details{
C function prototype: \code{CP_EXPORT cpBool cpSpaceContainsBody(cpSpace *space, cpBody *body);}
}
