% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpSpace.R
\name{cpSpaceContainsConstraint}
\alias{cpSpaceContainsConstraint}
\title{Test if a constraint has been added to the space.}
\usage{
cpSpaceContainsConstraint(space, constraint)
}
\arguments{
\item{space}{[\code{cpSpace *}]}

\item{constraint}{[\code{cpConstraint *}]}
}
\value{
[\code{cpBool}]
}
\description{
Test if a constraint has been added to the space.
}
\details{
C function prototype: \code{CP_EXPORT cpBool cpSpaceContainsConstraint(cpSpace *space, cpConstraint *constraint);}
}
