% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpSpace.R
\name{cpSpaceContainsShape}
\alias{cpSpaceContainsShape}
\title{Test if a collision shape has been added to the space.}
\usage{
cpSpaceContainsShape(space, shape)
}
\arguments{
\item{space}{[\code{cpSpace *}]}

\item{shape}{[\code{cpShape *}]}
}
\value{
[\code{cpBool}]
}
\description{
Test if a collision shape has been added to the space.
}
\details{
C function prototype: \code{CP_EXPORT cpBool cpSpaceContainsShape(cpSpace *space, cpShape *shape);}
}
