% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpSpace.R
\name{cpSpaceGetCurrentTimeStep}
\alias{cpSpaceGetCurrentTimeStep}
\title{Returns the current (or most recent) time step used with the given space. Useful
from callbacks if your time step is not a compile-time global.}
\usage{
cpSpaceGetCurrentTimeStep(space)
}
\arguments{
\item{space}{[\code{cpSpace *}]}
}
\value{
[\code{cpFloat}]
}
\description{
Returns the current (or most recent) time step used with the given space. Useful
from callbacks if your time step is not a compile-time global.
}
\details{
C function prototype: \code{CP_EXPORT cpFloat cpSpaceGetCurrentTimeStep(const cpSpace *space);}
}
