% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpSpace.R
\name{cpSpaceGetStaticBody}
\alias{cpSpaceGetStaticBody}
\title{The Space provided static body for a given cpSpace. This is merely provided for
convenience and you are not required to use it.}
\usage{
cpSpaceGetStaticBody(space)
}
\arguments{
\item{space}{[\code{cpSpace *}]}
}
\value{
[\code{cpBody *}]
}
\description{
The Space provided static body for a given cpSpace. This is merely provided for
convenience and you are not required to use it.
}
\details{
C function prototype: \code{CP_EXPORT cpBody* cpSpaceGetStaticBody(const cpSpace *space);}
}
