% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpSpace.R
\name{cpSpaceNew}
\alias{cpSpaceNew}
\title{Allocate and initialize a cpSpace.}
\usage{
cpSpaceNew()
}
\value{
[\code{cpSpace *}]
}
\description{
Allocate and initialize a cpSpace.
}
\details{
C function prototype: \code{CP_EXPORT cpSpace* cpSpaceNew(void);}
}
