% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpSpace.R
\name{cpSpaceReindexShape}
\alias{cpSpaceReindexShape}
\title{Update the collision detection data for a specific shape in the space.}
\usage{
cpSpaceReindexShape(space, shape)
}
\arguments{
\item{space}{[\code{cpSpace *}]}

\item{shape}{[\code{cpShape *}]}
}
\description{
Update the collision detection data for a specific shape in the space.
}
\details{
C function prototype: \code{CP_EXPORT void cpSpaceReindexShape(cpSpace *space, cpShape *shape);}
}
