% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpSpace.R
\name{cpSpaceReindexShapesForBody}
\alias{cpSpaceReindexShapesForBody}
\title{Update the collision detection data for all shapes attached to a body.}
\usage{
cpSpaceReindexShapesForBody(space, body)
}
\arguments{
\item{space}{[\code{cpSpace *}]}

\item{body}{[\code{cpBody *}]}
}
\description{
Update the collision detection data for all shapes attached to a body.
}
\details{
C function prototype: \code{CP_EXPORT void cpSpaceReindexShapesForBody(cpSpace *space, cpBody *body);}
}
