% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpSpace.R
\name{cpSpaceReindexStatic}
\alias{cpSpaceReindexStatic}
\title{Update the collision detection info for the static shapes in the space.}
\usage{
cpSpaceReindexStatic(space)
}
\arguments{
\item{space}{[\code{cpSpace *}]}
}
\description{
Update the collision detection info for the static shapes in the space.
}
\details{
C function prototype: \code{CP_EXPORT void cpSpaceReindexStatic(cpSpace *space);}
}
