% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpSpace.R
\name{cpSpaceRemoveBody}
\alias{cpSpaceRemoveBody}
\title{Remove a rigid body from the simulation.}
\usage{
cpSpaceRemoveBody(space, body)
}
\arguments{
\item{space}{[\code{cpSpace *}]}

\item{body}{[\code{cpBody *}]}
}
\description{
Remove a rigid body from the simulation.
}
\details{
C function prototype: \code{CP_EXPORT void cpSpaceRemoveBody(cpSpace *space, cpBody *body);}
}
