% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpSpace.R
\name{cpSpaceRemoveConstraint}
\alias{cpSpaceRemoveConstraint}
\title{Remove a constraint from the simulation.}
\usage{
cpSpaceRemoveConstraint(space, constraint)
}
\arguments{
\item{space}{[\code{cpSpace *}]}

\item{constraint}{[\code{cpConstraint *}]}
}
\description{
Remove a constraint from the simulation.
}
\details{
C function prototype: \code{CP_EXPORT void cpSpaceRemoveConstraint(cpSpace *space, cpConstraint *constraint);}
}
