% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpSpace.R
\name{cpSpaceRemoveShape}
\alias{cpSpaceRemoveShape}
\title{Remove a collision shape from the simulation.}
\usage{
cpSpaceRemoveShape(space, shape)
}
\arguments{
\item{space}{[\code{cpSpace *}]}

\item{shape}{[\code{cpShape *}]}
}
\description{
Remove a collision shape from the simulation.
}
\details{
C function prototype: \code{CP_EXPORT void cpSpaceRemoveShape(cpSpace *space, cpShape *shape);}
}
