% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpSpace.R
\name{cpSpaceSetCollisionBias}
\alias{cpSpaceSetCollisionBias}
\title{Determines how fast overlapping shapes are pushed apart. Expressed as a fraction
of the error remaining after each second. Defaults to pow(1.0 - 0.1, 60.0)
meaning that Chipmunk fixes 10\% of overlap each frame at 60Hz.}
\usage{
cpSpaceSetCollisionBias(space, collisionBias)
}
\arguments{
\item{space}{[\code{cpSpace *}]}

\item{collisionBias}{[\code{cpFloat}]}
}
\description{
Determines how fast overlapping shapes are pushed apart. Expressed as a fraction
of the error remaining after each second. Defaults to pow(1.0 - 0.1, 60.0)
meaning that Chipmunk fixes 10\% of overlap each frame at 60Hz.
}
\details{
C function prototype: \code{CP_EXPORT void cpSpaceSetCollisionBias(cpSpace *space, cpFloat collisionBias);}
}
