% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpSpace.R
\name{cpSpaceSetCollisionSlop}
\alias{cpSpaceSetCollisionSlop}
\title{Amount of encouraged penetration between colliding shapes. Used to reduce
oscillating contacts and keep the collision cache warm. Defaults to 0.1. If you
have poor simulation quality, increase this number as much as possible without
allowing visible amounts of overlap.}
\usage{
cpSpaceSetCollisionSlop(space, collisionSlop)
}
\arguments{
\item{space}{[\code{cpSpace *}]}

\item{collisionSlop}{[\code{cpFloat}]}
}
\description{
Amount of encouraged penetration between colliding shapes. Used to reduce
oscillating contacts and keep the collision cache warm. Defaults to 0.1. If you
have poor simulation quality, increase this number as much as possible without
allowing visible amounts of overlap.
}
\details{
C function prototype: \code{CP_EXPORT void cpSpaceSetCollisionSlop(cpSpace *space, cpFloat collisionSlop);}
}
