% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpSpace.R
\name{cpSpaceSetDamping}
\alias{cpSpaceSetDamping}
\title{Damping rate expressed as the fraction of velocity bodies retain each second. A
value of 0.9 would mean that each body's velocity will drop 10\% per second. The
default value is 1.0, meaning no damping is applied. @note This damping value is
different than those of cpDampedSpring and cpDampedRotarySpring.}
\usage{
cpSpaceSetDamping(space, damping)
}
\arguments{
\item{space}{[\code{cpSpace *}]}

\item{damping}{[\code{cpFloat}]}
}
\description{
Damping rate expressed as the fraction of velocity bodies retain each second. A
value of 0.9 would mean that each body's velocity will drop 10\% per second. The
default value is 1.0, meaning no damping is applied. @note This damping value is
different than those of cpDampedSpring and cpDampedRotarySpring.
}
\details{
C function prototype: \code{CP_EXPORT void cpSpaceSetDamping(cpSpace *space, cpFloat damping);}
}
