% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpSpace.R
\name{cpSpaceSetGravity}
\alias{cpSpaceSetGravity}
\title{Gravity to pass to rigid bodies when integrating velocity.}
\usage{
cpSpaceSetGravity(space, gravity)
}
\arguments{
\item{space}{[\code{cpSpace *}]}

\item{gravity}{[\code{cpVect *}]}
}
\description{
Gravity to pass to rigid bodies when integrating velocity.
}
\details{
C function prototype: \code{CP_EXPORT void cpSpaceSetGravity(cpSpace *space, cpVect gravity);}
}
