% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpSpace.R
\name{cpSpaceSetIdleSpeedThreshold}
\alias{cpSpaceSetIdleSpeedThreshold}
\title{Speed threshold for a body to be considered idle. The default value of 0 means
to let the space guess a good threshold based on gravity.}
\usage{
cpSpaceSetIdleSpeedThreshold(space, idleSpeedThreshold)
}
\arguments{
\item{space}{[\code{cpSpace *}]}

\item{idleSpeedThreshold}{[\code{cpFloat}]}
}
\description{
Speed threshold for a body to be considered idle. The default value of 0 means
to let the space guess a good threshold based on gravity.
}
\details{
C function prototype: \code{CP_EXPORT void cpSpaceSetIdleSpeedThreshold(cpSpace *space, cpFloat idleSpeedThreshold);}
}
