% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpSpace.R
\name{cpSpaceSetIterations}
\alias{cpSpaceSetIterations}
\title{Number of iterations to use in the impulse solver to solve contacts and other
constraints.}
\usage{
cpSpaceSetIterations(space, iterations)
}
\arguments{
\item{space}{[\code{cpSpace *}]}

\item{iterations}{[\code{int}]}
}
\description{
Number of iterations to use in the impulse solver to solve contacts and other
constraints.
}
\details{
C function prototype: \code{CP_EXPORT void cpSpaceSetIterations(cpSpace *space, int iterations);}
}
