% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpSpace.R
\name{cpSpaceSetSleepTimeThreshold}
\alias{cpSpaceSetSleepTimeThreshold}
\title{Time a group of bodies must remain idle in order to fall asleep. Enabling
sleeping also implicitly enables the the contact graph. The default value of
INFINITY disables the sleeping algorithm.}
\usage{
cpSpaceSetSleepTimeThreshold(space, sleepTimeThreshold)
}
\arguments{
\item{space}{[\code{cpSpace *}]}

\item{sleepTimeThreshold}{[\code{cpFloat}]}
}
\description{
Time a group of bodies must remain idle in order to fall asleep. Enabling
sleeping also implicitly enables the the contact graph. The default value of
INFINITY disables the sleeping algorithm.
}
\details{
C function prototype: \code{CP_EXPORT void cpSpaceSetSleepTimeThreshold(cpSpace *space, cpFloat sleepTimeThreshold);}
}
