% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpSpace.R
\name{cpSpaceStep}
\alias{cpSpaceStep}
\title{Step the space forward in time by @c dt.}
\usage{
cpSpaceStep(space, dt)
}
\arguments{
\item{space}{[\code{cpSpace *}]}

\item{dt}{[\code{cpFloat}]}
}
\description{
Step the space forward in time by @c dt.
}
\details{
C function prototype: \code{CP_EXPORT void cpSpaceStep(cpSpace *space, cpFloat dt);}
}
