% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpTransform.R
\name{cpTransformMult}
\alias{cpTransformMult}
\title{Multiply two transformation matrices.}
\usage{
cpTransformMult(t1, t2)
}
\arguments{
\item{t1}{[\code{cpTransform *}]}

\item{t2}{[\code{cpTransform *}]}
}
\value{
[\code{cpTransform *}]
}
\description{
Multiply two transformation matrices.
}
\details{
C function prototype: \code{static inline cpTransform cpTransformMult(cpTransform t1, cpTransform t2)}
}
