% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpTransform.R
\name{cpTransformNew}
\alias{cpTransformNew}
\title{Construct a new transform matrix. (a, b) is the x basis vector. (c, d) is the y
basis vector. (tx, ty) is the translation.}
\usage{
cpTransformNew(a, b, c, d, tx, ty)
}
\arguments{
\item{a}{[\code{cpFloat}]}

\item{b}{[\code{cpFloat}]}

\item{c}{[\code{cpFloat}]}

\item{d}{[\code{cpFloat}]}

\item{tx}{[\code{cpFloat}]}

\item{ty}{[\code{cpFloat}]}
}
\value{
[\code{cpTransform *}]
}
\description{
Construct a new transform matrix. (a, b) is the x basis vector. (c, d) is the y
basis vector. (tx, ty) is the translation.
}
\details{
C function prototype: \code{static inline cpTransform cpTransformNew(cpFloat a, cpFloat b, cpFloat c, cpFloat d, cpFloat tx, cpFloat ty)}
}
