% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpTransform.R
\name{cpTransformRigid}
\alias{cpTransformRigid}
\title{Create a rigid transformation matrix. (transation + rotation)}
\usage{
cpTransformRigid(translate, radians)
}
\arguments{
\item{translate}{[\code{cpVect *}]}

\item{radians}{[\code{cpFloat}]}
}
\value{
[\code{cpTransform *}]
}
\description{
Create a rigid transformation matrix. (transation + rotation)
}
\details{
C function prototype: \code{static inline cpTransform cpTransformRigid(cpVect translate, cpFloat radians)}
}
