% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpTransform.R
\name{cpTransformRotate}
\alias{cpTransformRotate}
\title{Create a rotation matrix.}
\usage{
cpTransformRotate(radians)
}
\arguments{
\item{radians}{[\code{cpFloat}]}
}
\value{
[\code{cpTransform *}]
}
\description{
Create a rotation matrix.
}
\details{
C function prototype: \code{static inline cpTransform cpTransformRotate(cpFloat radians)}
}
