% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpTransform.R
\name{cpTransformScale}
\alias{cpTransformScale}
\title{Create a scale matrix.}
\usage{
cpTransformScale(scaleX, scaleY)
}
\arguments{
\item{scaleX}{[\code{cpFloat}]}

\item{scaleY}{[\code{cpFloat}]}
}
\value{
[\code{cpTransform *}]
}
\description{
Create a scale matrix.
}
\details{
C function prototype: \code{static inline cpTransform cpTransformScale(cpFloat scaleX, cpFloat scaleY)}
}
