% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpTransform.R
\name{cpTransformTranslate}
\alias{cpTransformTranslate}
\title{Create a transation matrix.}
\usage{
cpTransformTranslate(translate)
}
\arguments{
\item{translate}{[\code{cpVect *}]}
}
\value{
[\code{cpTransform *}]
}
\description{
Create a transation matrix.
}
\details{
C function prototype: \code{static inline cpTransform cpTransformTranslate(cpVect translate)}
}
