% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpVect-structs.R
\name{cpVect}
\alias{cpVect}
\alias{cpv}
\title{Create a pointer to a \code{cpVect} struct}
\usage{
cpVect(x, y)

cpv(x, y)
}
\arguments{
\item{x, y}{[double] coordiantes. vectorised}
}
\description{
Create a pointer to a \code{cpVect} struct
}
