% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpBody-bespoke.R
\name{get_all_body_info}
\alias{get_all_body_info}
\title{Get all info for a list of bodies}
\usage{
get_all_body_info(bodies)
}
\arguments{
\item{bodies}{list of \code{cpBody} objects}
}
\value{
data.frame
}
\description{
Get all info for a list of bodies
}
