% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpBody-bespoke.R
\name{get_body_state}
\alias{get_body_state}
\title{Get the current state for every body in a list}
\usage{
get_body_state(bodies)
}
\arguments{
\item{bodies}{list of \code{cpBody} objects}
}
\value{
data.frame
}
\description{
Get the current state for every body in a list
}
