% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6.R
\name{Chipmunk}
\alias{Chipmunk}
\title{Chipmunk class}
\description{
Chipmunk class

Chipmunk class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Chipmunk$new()}}
\item \href{#method-add_segment}{\code{Chipmunk$add_segment()}}
\item \href{#method-get_segments}{\code{Chipmunk$get_segments()}}
\item \href{#method-add_body}{\code{Chipmunk$add_body()}}
\item \href{#method-get_bodies}{\code{Chipmunk$get_bodies()}}
\item \href{#method-remove_body}{\code{Chipmunk$remove_body()}}
\item \href{#method-advance}{\code{Chipmunk$advance()}}
\item \href{#method-clone}{\code{Chipmunk$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize the simulation space
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Chipmunk$new(gravity = cpv(0, -100), time_step = 0.01)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{gravity}}{vector object indication direction and strength
of gravity. Default: \code{cpv(0, -100)}}

\item{\code{time_step}}{simulation step size. smaller is more accurate.
       Default: 0.01
Add a static segment}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_segment"></a>}}
\if{latex}{\out{\hypertarget{method-add_segment}{}}}
\subsection{Method \code{add_segment()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Chipmunk$add_segment(x1, y1, x2, y2, friction = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x1, y1, x2, y2}}{segment end points}

\item{\code{friction}}{friction along ground. default 1.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_segments"></a>}}
\if{latex}{\out{\hypertarget{method-get_segments}{}}}
\subsection{Method \code{get_segments()}}{
Get the data.frame of all current segments
Add a body to the space. Currently only supports circular bodies.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Chipmunk$get_segments()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_body"></a>}}
\if{latex}{\out{\hypertarget{method-add_body}{}}}
\subsection{Method \code{add_body()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Chipmunk$add_body(
  x,
  y,
  vx = 0,
  vy = 0,
  radius = 1,
  mass = 1,
  friction = 0.7,
  elasticity = 0
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x, y}}{initial body location}

\item{\code{vx, vy}}{initial body velocity}

\item{\code{radius}}{radius of body. default: 1}

\item{\code{mass}}{mass of body. default: 1}

\item{\code{friction}}{default: 0.7}

\item{\code{elasticity}}{default: 0  (no bounce).  Valid range [0, 1]}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_bodies"></a>}}
\if{latex}{\out{\hypertarget{method-get_bodies}{}}}
\subsection{Method \code{get_bodies()}}{
Get the location of all the bodies
Remove a body from the simulation

To speed up the simulation process, you could test bodies for
location and remove those you are no longer interested in.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Chipmunk$get_bodies()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-remove_body"></a>}}
\if{latex}{\out{\hypertarget{method-remove_body}{}}}
\subsection{Method \code{remove_body()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Chipmunk$remove_body(i)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{i}}{index of body
Advance the simulation by N timesteps}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-advance"></a>}}
\if{latex}{\out{\hypertarget{method-advance}{}}}
\subsection{Method \code{advance()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Chipmunk$advance(N = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{N}}{number of time steps to advance}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Chipmunk$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
