% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpShape.R
\name{cpCircleShapeNew}
\alias{cpCircleShapeNew}
\title{cpCircleShapeNew}
\usage{
cpCircleShapeNew(body, radius, offset)
}
\arguments{
\item{body}{\code{cpBody *}}

\item{radius}{\code{cpFloat}}

\item{offset}{\code{cpVect *}}
}
\description{
\code{ CP_EXPORT cpShape* cpCircleShapeNew(cpBody *body, cpFloat radius, cpVect offset); }
}
