% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpShape.R
\name{cpShapeSetFriction}
\alias{cpShapeSetFriction}
\title{cpShapeSetFriction}
\usage{
cpShapeSetFriction(shape, friction)
}
\arguments{
\item{shape}{\code{cpShape *}}

\item{friction}{\code{cpFloat}}
}
\description{
\code{ CP_EXPORT void cpShapeSetFriction(cpShape *shape, cpFloat friction); }
}
